+++ worksheets/worksheet05/size5.c
@@ -1,7 +1,7 @@
 #include <stdio.h> 
-int global = 10; /* Initialized global variable stored in DS */ 
+long global = 10; /* Initialized global variable stored in DS */ 
 int main(void)
 {
-	static int i = 100; /* Initialized static variable stored in DS */ 
+	static long i = 100; /* Initialized static variable stored in DS */ 
 	return 0; 
 }
